payment DTO and kafka listern updated in wallet service

+++ transaction-service/src/main/java/com/paypal/transaction_service/dtos/PaymentRequestDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Builder
@@ -14,5 +15,5 @@ import java.math.BigInteger;
 public class PaymentRequestDto {
     private String senderId;
     private String receiverId;
-    private BigInteger amount;
+    private BigDecimal amount;
 }
